Define the store for the bus stops
It is an implementation of Map, to be able to perform map operations on the store. The Key will be the bus stop identifier, represented by the String "P01" i.e.

Besides, it has a static map to hold references to bus stops, as they are limited to a small set of elements (we know all bus stops in all routes), so we will keep all bus stops in memory, which will faster operations.